When multiple characters are entered!
Hey, I found that when user entered multiple letters instead of one letter, The program will consider the whole thing and will consider it as wrong. With this simple edit it will only select the first character if multiple characters are entered. 
Example:
if "aedfdfdf" is entered by the user, the program will only take 'a' into consideration and will treat as user entered 'a'. 

This is not ideal but will make the program a little bit more kinda smooth. I suggest you to make the program re prompt the user to input a single character instead of multiple characters. Thank you!